+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 7.82
 -----
-
+*   Bug Fixes
+    *   Fix status bar theming during onboarding.
+        ([#3460](https://github.com/Automattic/pocket-casts-android/pull/3460))
+*   Updates
+    *   Do not close episode details screen when archiving an episode
+        ([#3473](https://github.com/Automattic/pocket-casts-android/pull/3473))
 
 7.81
 -----